# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee Rating OpenApi Documentation",
        "description": "Employee Rating OpenApi description",
        "contact": {
            "email": "info@employeerating.live"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "EMPLOYEERATING"
        }
    ],
    "paths": {
...
        "/vote_now": {
            "post": {
                "tags": [
                    "Votting Management"
                ],
                "summary": "Do Vote",
                "operationId": "voteNow",
                "parameters": [
                    {
                        "name": "candidate_id",
                        "in": "query",
                        "description": "Enter candidate id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "technical_skill",
                        "in": "query",
                        "description": "Enter rating for technical skill (scale 1-10)",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "attitude",
                        "in": "query",
                        "description": "Enter rating for attitude (scale 1-10)",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "problem_solving_skill",
                        "in": "query",
                        "description": "Enter rating for problem solving skill (scale 1-10)",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}